Extract stuff specific to regen-conf from service.py to regenconf.py
